move "basic" builtins to TableGen
This is the first step towards a move of all builtin function
definitions living in TableGen.  The intent is that once that is
accomplished, the storage cost and runtime cost (due to relocations) of
builtin function info structures can be significantly reduced by using
integer offsets into TableGen'd string tables rather than pointers.

We need to add an additional API to TableGen, as uses of at least the
atomic builtins depend on the order in which they are specified in
Builtins.def/Builtins.td, and changing `getDerivedDefinitions` to emit
records in definition order had other undesirable effects that I was not
prepared to untangle.

The TableGen definitions live in a separate file to make future
conversions of the architecture-specific builtins more straightforward.

No functional change intended.